reformated the code of some files.
