+++ src/services/File Management/package.json
@@ -2,9 +2,10 @@
   "name": "file-management",
   "version": "1.0.0",
   "description": "service to store files",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "nodemon server.js",
+    "prod": "node server.js"
   },
   "author": "",
   "license": "ISC",
