+++ src/services/File Management/rmq/rmqConsumerServer.js
@@ -1,15 +1,14 @@
 const path = require("path");
 const amqp = require("amqplib");
 const fs = require("fs");
-const { PrismaClient } = require("@prisma/client");
-const prisma = new PrismaClient();
+const prisma = require("../prisma/prismaClient");
 const uploadDir = path.join(__dirname, "../uploads");
 const rabbitMqUrl = process.env.RMQURL;
 async function saveFileToDisk(fileBuffer, filename, fileId) {
   try {
     const filePath = path.join(uploadDir, filename);
     fs.writeFileSync(filePath, fileBuffer);
-    console.log(`File saved to disk: ${filePath}`);
+    console.log(`[!] File saved to disk: ${filePath}`);
     await prisma.file.update({
       where: {
         id: fileId,
@@ -34,7 +33,7 @@ async function consume() {
   const connection = await amqp.connect(rabbitMqUrl);
   const channel = await connection.createChannel();
   await channel.assertQueue("files", { durable: true });
-  console.log("Waiting for messages in queue...");
+  console.log("[!] Waiting for messages in queue...");
   channel.consume("files", async (message) => {
     if (message !== null) {
       const content = JSON.parse(message.content.toString());
@@ -54,4 +53,4 @@ async function consume() {
   });
 }
 
-consume();
+module.exports = consume;
